zeliaoms/requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap5==2025.4
data==0.4
decorator==5.2.1
Django==5.2.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
dnspython==2.7.0
et_xmlfile==2.0.0
funcsigs==1.0.2
future==1.0.0
latex==0.7.0

mysql-connector==2.2.9
mysql-connector-python==9.3.0


openpyxl==3.1.5
pymysql
djangorestframework
pillow>=11.3.0

python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
shutilwhich==1.1.0
six==1.17.0
sqlparse==0.5.3
tempdir==0.7.1
tzdata==2025.2
whitenoise
